LPD-90353 Collapse Input and Display duplication into a shared module
TagInput.tsx and VocabularyInput.tsx duplicated several hundred lines: the EVENT_TYPE_OPTIONS, OCCURRENCE_OPTIONS, and ASSET_TYPE_OPTIONS arrays, the DEFAULT_CONJUNCTION_CRITERION constant, the isValidOccurrenceCount helper, and three identical readers (getAssetTypeFromValue, getEventTypeFromValue, getConjunctionCriterionFromValue), plus a near-identical buildValue whose only differences were the OData property names and whether categories were included. inputs/shared/remote-filter-input-helpers.ts now owns all of that, along with a parameterized buildRemoteFilterValue. Each Input becomes a thin wrapper that supplies its config (idProperty, nameProperty, supportsCategories) and re-exports the helpers, keeping the tests that import buildValue and the constants from the Input modules unchanged.

VocabularyDisplay.tsx and TagDisplay.tsx no longer reach across to the sibling Input file to obtain their shared constants; they import directly from the same shared helpers module.
